requirements.txt
chromadb-client>=1.5.0
sentence-transformers>=3.0.0
